Change sqrt(atoi... to sqrt(atof... to get around a PGI compiler error.
Former-commit-id: e11c33446784594c8c8d7d641d81569d61d6ed79

+++ src/tools/core/stat_analysis/aggr_stat_line.cc
@@ -347,7 +347,7 @@ StatHdrColumns StatHdrInfo::get_shc(const ConcatString &cur_case,
    }
    
    if(hdr_name.has("INTERP_PNTS", index)) {
-      wdth = nint(sqrt(atoi(hdr_value[index])));
+      wdth = nint(sqrt(atof(hdr_value[index])));
    }
    shc.set_interp_wdth(wdth);
